Add files via upload
This application provides a reinforcement solution for power control in 5G networks in order to achieve best Quality of Service.In particular, collects average SINR from the bandwidth parts of a gNodeB  and stores them into a dataset and applies an action based on transmission power through  an agent. The agent choose the action via epsilon-greedy srategy  and rewards the network based on signal quality.